🧪 : axios로 인스턴한 get mocking페이지에서 적용 및 테스트
